diy_smithing: armor repair
allow the moderately lossy repair of armor, representing the removal of
bad parts and replacement with good parts from other armor pieces. The
new armor is just as functional, but loses 50% of its value and is likely
to be rough on top of that.

lamellar requires 115% of the final item's weight in lamellar scraps,
20% of the final item's weight in leather, and a knife.
the tribesman cuts out bad lames and uses the leather to link them back
into the item.

mail requires 110% of the item's final weight in mail, 10% of item's final
weight in cloth, a knife, and an iron hammer and tongs. Very large items
also require a pound of wrought iron.
the tribesemen cuts out bad links and padding and sews in new padding and
uses the hammer and tongs to rivet in new links.

iron pieces require 100% of the item's final weight in iron, 50% of the
item's final weight in wrough iron, and a full forge rig including charcoal
and a fire.
the tribesmen welds/rivets in patches.